7_El3yada_Postgres_Sql/requirements.txt
matplotlib
numpy
pandas
plotly
python-dotenv
PyYAML
streamlit
streamlit_authenticator
streamlit_extras
streamlit_option_menu
streamlit_phone_number
psycopg2-binary
